Refactor OpenTelemetry integration and enhance event handling
- Updated the `OtelEnvelope` structure to represent a batch of OpenTelemetry log/event records, replacing the previous attributes and events model.
- Introduced new `OtelEvent` and `OtelAttribute` structures to better align with the OTLP log data model, including additional fields for enhanced event metadata.
- Refactored the `to_otel_envelope` method to generate records instead of separate attributes and events, improving clarity and usability.
- Adjusted related tests and documentation to reflect the new structure and ensure accurate behavior in OpenTelemetry data handling.

+++ diagweave/src/report.rs
@@ -348,7 +348,7 @@ impl<E> Report<E> {
 
     /// Sets the trace flags.
     #[cfg(feature = "trace")]
-    pub fn with_trace_flags(mut self, flags: u32) -> Self {
+    pub fn with_trace_flags(mut self, flags: u8) -> Self {
         let trace = self.trace_mut();
         trace.context.flags = Some(flags);
         sync_trace_sampled_with_flags(&mut trace.context);
